Separate routines into validator-full.js and validator-light.js
