+++ validator/validator.js
@@ -15,11 +15,7 @@
  * limitations under the license.
  */
 goog.provide('amp.validator.CssLengthAndUnit');  // Only for testing.
-goog.provide('amp.validator.Terminal');
-goog.provide('amp.validator.annotateWithErrorCategories');
 goog.provide('amp.validator.isSeverityWarning');
-goog.provide('amp.validator.renderErrorMessage');
-goog.provide('amp.validator.renderValidationResult');
 goog.provide('amp.validator.validateString');
 goog.require('amp.htmlparser.HtmlParser');
 goog.require('amp.htmlparser.HtmlSaxHandlerWithLocation');
@@ -35,13 +31,13 @@ goog.require('amp.validator.ValidationError.Severity');
 goog.require('amp.validator.ValidationResult');
 goog.require('amp.validator.ValidationResult.Status');
 goog.require('amp.validator.ValidatorRules');
+goog.require('amp.validator.maxSpecificity');
 goog.require('goog.Uri');
 goog.require('goog.array');
 goog.require('goog.asserts');
 goog.require('goog.string');
 goog.require('goog.structs.Map');
 goog.require('goog.structs.Set');
-goog.require('goog.uri.utils');
 goog.require('parse_css.BlockType');
 goog.require('parse_css.ParsedCssUrl');
 goog.require('parse_css.RuleVisitor');
@@ -109,163 +105,6 @@ function subtractDiff(left, right) {
 }
 
 /**
- * A higher number means |code| is more specific, meaning more
- * helpful. Used by maxSpecificity below.
- * @param {!amp.validator.ValidationError.Code} code
- * @return {number}
- */
-function specificity(code) {
-  switch (code) {
-    case amp.validator.ValidationError.Code.UNKNOWN_CODE:
-      return 0;
-    case amp.validator.ValidationError.Code
-        .MANDATORY_CDATA_MISSING_OR_INCORRECT:
-      return 1;
-    case amp.validator.ValidationError.Code.CDATA_VIOLATES_BLACKLIST:
-      return 2;
-    case amp.validator.ValidationError.Code.WRONG_PARENT_TAG:
-      return 3;
-    case amp.validator.ValidationError.Code.DISALLOWED_TAG_ANCESTOR:
-      return 4;
-    case amp.validator.ValidationError.Code.MANDATORY_TAG_ANCESTOR:
-      return 5;
-    case amp.validator.ValidationError.Code.MANDATORY_TAG_ANCESTOR_WITH_HINT:
-      return 6;
-    case amp.validator.ValidationError.Code.MANDATORY_TAG_MISSING:
-      return 7;
-    case amp.validator.ValidationError.Code.TAG_REQUIRED_BY_MISSING:
-      return 8;
-    case amp.validator.ValidationError.Code.ATTR_REQUIRED_BUT_MISSING:
-      return 9;
-    case amp.validator.ValidationError.Code.DISALLOWED_TAG:
-      return 10;
-    case amp.validator.ValidationError.Code.DISALLOWED_ATTR:
-      return 11;
-    case amp.validator.ValidationError.Code.INVALID_ATTR_VALUE:
-      return 12;
-    case amp.validator.ValidationError.Code.ATTR_VALUE_REQUIRED_BY_LAYOUT:
-      return 13;
-    case amp.validator.ValidationError.Code.MANDATORY_ATTR_MISSING:
-      return 14;
-    case amp.validator.ValidationError.Code.MANDATORY_ONEOF_ATTR_MISSING:
-      return 15;
-    case amp.validator.ValidationError.Code.DUPLICATE_UNIQUE_TAG:
-      return 16;
-    case amp.validator.ValidationError.Code.STYLESHEET_TOO_LONG_OLD_VARIANT:
-      return 17;
-    case amp.validator.ValidationError.Code.STYLESHEET_TOO_LONG:
-      return 18;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX:
-      return 19;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_AT_RULE:
-      return 20;
-    case amp.validator.ValidationError.Code
-        .MANDATORY_PROPERTY_MISSING_FROM_ATTR_VALUE:
-      return 21;
-    case amp.validator.ValidationError.Code
-        .INVALID_PROPERTY_VALUE_IN_ATTR_VALUE:
-      return 22;
-    case amp.validator.ValidationError.Code.DISALLOWED_PROPERTY_IN_ATTR_VALUE:
-      return 23;
-    case amp.validator.ValidationError.Code.MUTUALLY_EXCLUSIVE_ATTRS:
-      return 24;
-    case amp.validator.ValidationError.Code.UNESCAPED_TEMPLATE_IN_ATTR_VALUE:
-      return 25;
-    case amp.validator.ValidationError.Code.TEMPLATE_PARTIAL_IN_ATTR_VALUE:
-      return 26;
-    case amp.validator.ValidationError.Code.TEMPLATE_IN_ATTR_NAME:
-      return 27;
-    case amp.validator.ValidationError.Code
-        .INCONSISTENT_UNITS_FOR_WIDTH_AND_HEIGHT:
-      return 28;
-    case amp.validator.ValidationError.Code.IMPLIED_LAYOUT_INVALID:
-      return 29;
-    case amp.validator.ValidationError.Code.SPECIFIED_LAYOUT_INVALID:
-      return 30;
-    case amp.validator.ValidationError.Code.DEV_MODE_ENABLED:
-      return 31;
-    case amp.validator.ValidationError.Code.ATTR_DISALLOWED_BY_IMPLIED_LAYOUT:
-      return 32;
-    case amp.validator.ValidationError.Code.ATTR_DISALLOWED_BY_SPECIFIED_LAYOUT:
-      return 33;
-    case amp.validator.ValidationError.Code.DUPLICATE_DIMENSION:
-      return 34;
-    case amp.validator.ValidationError.Code.DISALLOWED_RELATIVE_URL:
-      return 35;
-    case amp.validator.ValidationError.Code.MISSING_URL:
-      return 36;
-    case amp.validator.ValidationError.Code.INVALID_URL_PROTOCOL:
-      return 37;
-    case amp.validator.ValidationError.Code.INVALID_URL:
-      return 38;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_STRAY_TRAILING_BACKSLASH:
-      return 39;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_UNTERMINATED_COMMENT:
-      return 40;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_UNTERMINATED_STRING:
-      return 41;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_BAD_URL:
-      return 42;
-    case amp.validator.ValidationError.Code
-        .CSS_SYNTAX_EOF_IN_PRELUDE_OF_QUALIFIED_RULE:
-      return 43;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_DECLARATION:
-      return 44;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_INCOMPLETE_DECLARATION:
-      return 45;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_ERROR_IN_PSEUDO_SELECTOR:
-      return 46;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_MISSING_SELECTOR:
-      return 47;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_NOT_A_SELECTOR_START:
-      return 48;
-    case amp.validator.ValidationError.Code
-        .CSS_SYNTAX_UNPARSED_INPUT_REMAINS_IN_SELECTOR:
-      return 49;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_MISSING_URL:
-      return 50;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_URL:
-      return 51;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_URL_PROTOCOL:
-      return 52;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_DISALLOWED_RELATIVE_URL:
-      return 53;
-    case amp.validator.ValidationError.Code.INCORRECT_NUM_CHILD_TAGS:
-      return 54;
-    case amp.validator.ValidationError.Code.DISALLOWED_CHILD_TAG_NAME:
-      return 55;
-    case amp.validator.ValidationError.Code.DISALLOWED_FIRST_CHILD_TAG_NAME:
-      return 56;
-    case amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_ATTR_SELECTOR:
-      return 57;
-    case amp.validator.ValidationError.Code.GENERAL_DISALLOWED_TAG:
-      return 100;
-    case amp.validator.ValidationError.Code.DEPRECATED_ATTR:
-      return 101;
-    case amp.validator.ValidationError.Code.DEPRECATED_TAG:
-      return 102;
-    default:
-      goog.asserts.fail('Unrecognized Code: ' + code);
-  }
-}
-
-/**
- * A helper function which allows us to compare two candidate results
- * in validateTag to report the results which have the most specific errors.
- * @param {!amp.validator.ValidationResult} validationResult
- * @return {number} maximum value of specificity found in all errors.
- */
-function maxSpecificity(validationResult) {
-  let max = 0;
-  for (const error of validationResult.errors) {
-    goog.asserts.assert(error.code != null);
-    const thisSpecificity = specificity(error.code);
-    max = Math.max(thisSpecificity, max);
-  }
-  return max;
-}
-
-/**
  * Merge results from another ValidationResult while dealing with the UNKNOWN
  *   status.
  * @param {!amp.validator.ValidationResult} other
@@ -277,110 +116,6 @@ amp.validator.ValidationResult.prototype.mergeFrom = function(other) {
     this.status = other.status;
   }
   Array.prototype.push.apply(this.errors, other.errors);
-};
-
-/**
- * The terminal is an abstraction for the window.console object which
- * accomodates differences between console implementations and provides
- * a convenient way to capture what's being emitted to the terminal
- * in a unittest. Pass the optional parameter to the constructor
- * to observe the calls that would have gone to window.console otherwise.
- * @constructor
- */
-amp.validator.Terminal = class {
-  /**
-   * @param {!Array<string>=} opt_out an array into which the terminal will
-   *     emit one string per info / warn / error calls.
-   */
-  constructor(opt_out) { this.out_ = opt_out || null; }
-
-  /** @param {!string} msg */
-  info(msg) {
-    if (this.out_) {
-      this.out_.push('I: ' + msg);
-    } else {
-      (console.info || console.log).call(console, msg);
-    }
-  }
-
-  /** @param {!string} msg */
-  warn(msg) {
-    if (this.out_) {
-      this.out_.push('W: ' + msg);
-    } else if (console.warn) {
-      console.warn(msg);
-    } else {
-      console.log('WARNING: ' + msg);
-    }
-  }
-
-  /** @param {!string} msg */
-  error(msg) {
-    if (this.out_) {
-      this.out_.push('E: ' + msg);
-    } else if (console.error) {
-      console.error(msg);
-    } else {
-      console.log('ERROR: ' + msg);
-    }
-  }
-}
-
-/**
- * Emits this validation result to the terminal, distinguishing warnings and
- *   errors.
- * @param {string} url
- * @param {!amp.validator.Terminal=} opt_terminal
- * @param {!string=} opt_errorCategoryFilter
- */
-amp.validator.ValidationResult.prototype.outputToTerminal = function(
-    url, opt_terminal, opt_errorCategoryFilter) {
-
-  const terminal = opt_terminal || new amp.validator.Terminal();
-  const errorCategoryFilter = opt_errorCategoryFilter || null;
-
-  const status = this.status;
-  if (status === amp.validator.ValidationResult.Status.PASS) {
-    terminal.info('AMP validation successful.');
-    return;
-  }
-  if (status !== amp.validator.ValidationResult.Status.FAIL) {
-    terminal.error(
-        'AMP validation had unknown results. This should not happen.');
-    return;
-  }
-  let errors;
-  if (errorCategoryFilter === null) {
-    terminal.error('AMP validation had errors:');
-    errors = this.errors;
-  } else {
-    errors = [];
-    for (const error of this.errors) {
-      if (amp.validator.categorizeError(error) === errorCategoryFilter) {
-        errors.push(error);
-      }
-    }
-    const urlWithoutFilter =
-        goog.uri.utils.removeFragment(url) + '#development=1';
-    if (errors.length === 0) {
-      terminal.error(
-          'AMP validation - no errors matching ' +
-          'filter=' + errorCategoryFilter + ' found. ' +
-          'To see all errors, visit ' + urlWithoutFilter);
-    } else {
-      terminal.error(
-          'AMP validation - displaying errors matching ' +
-          'filter=' + errorCategoryFilter + '. ' +
-          'To see all errors, visit ' + urlWithoutFilter);
-    }
-  }
-  for (const error of errors) {
-    if (error.severity === amp.validator.ValidationError.Severity.ERROR) {
-      terminal.error(errorLine(url, error));
-    } else {
-      terminal.warn(errorLine(url, error));
-    }
-  }
 };
 
 /**
@@ -2870,8 +2605,8 @@ class ParsedValidatorRules {
       }
 
       // If the same number of errors, prefer the most specific error.
-      if (maxSpecificity(resultForAttempt) >
-          maxSpecificity(resultForBestAttempt)) {
+      if (amp.validator.maxSpecificity(resultForAttempt) >
+          amp.validator.maxSpecificity(resultForBestAttempt)) {
         resultForBestAttempt.status = resultForAttempt.status;
         resultForBestAttempt.errors = resultForAttempt.errors;
       }
@@ -3205,395 +2940,6 @@ function applyFormat(format, error) {
   }
   return message.replace(new RegExp('%%', 'g'), '%');
 }
-
-/**
- * Renders the error message for a single error, regardless of whether
- * or not it has an associated format.
- * @param {!amp.validator.ValidationError} error
- * @return {!string}
- * @export
- */
-amp.validator.renderErrorMessage = function(error) {
-  let out = '';
-  const format =
-      parsedValidatorRulesSingleton.getFormatByCode().get(error.code);
-  // A11Y errors are special cased and don't have parameters.
-  if (format !== undefined && error.params.length > 0) {
-    out += applyFormat(format, error);
-  } else {
-    out += error.code;
-    if (error.detail !== undefined) out += ' ' + error.detail;
-  }
-  return out;
-};
-
-/**
- * Renders one line of error output.
- * @param {!string} filenameOrUrl
- * @param {!amp.validator.ValidationError} error
- * @return {!string}
- */
-function errorLine(filenameOrUrl, error) {
-  const line = error.line || 1;
-  const col = error.col || 0;
-
-  let errorLine = goog.uri.utils.removeFragment(filenameOrUrl) + ':' + line +
-      ':' + col + ' ';
-  errorLine += amp.validator.renderErrorMessage(error);
-  if (error.specUrl) {
-    errorLine += ' (see ' + error.specUrl + ')';
-  }
-  if (error.category !== null) {
-    errorLine += ' [' + error.category + ']';
-  }
-  return errorLine;
-}
-
-/**
- * Renders the validation results into an array of human readable strings.
- * Careful when modifying this - it's called from
- * https://github.com/ampproject/amphtml/blob/master/test/integration/test-example-validation.js.
- * @param {!Object} validationResult
- * @param {string} filename to use in rendering error messages.
- * @return {!Array<string>}
- * @export
- */
-amp.validator.renderValidationResult = function(validationResult, filename) {
-  const rendered = [];
-  rendered.push(validationResult.status);
-  for (const error of validationResult.errors) {
-    rendered.push(errorLine(filename, error));
-  }
-  return rendered;
-};
-
-/**
- * Computes the validation category for this |error|. This is a higher
- * level classification that distinguishes layout problems, problems
- * with specific tags, etc. The category is determined with heuristics,
- * just based on the information in |error|. We consider
- * ValidationError::Code, ValidationError::params (including suffix /
- * prefix matches.
- * @param {!amp.validator.ValidationError} error
- * @return {!amp.validator.ErrorCategory.Code}
- * @export
- */
-amp.validator.categorizeError = function(error) {
-  // This shouldn't happen in practice. We always set some params, and
-  // UNKNOWN_CODE would indicate that the field wasn't populated.
-  if (error.params.length === 0 ||
-      error.code === amp.validator.ValidationError.Code.UNKNOWN_CODE ||
-      error.code === null) {
-    return amp.validator.ErrorCategory.Code.UNKNOWN;
-  }
-  // E.g. "The tag 'img' may only appear as a descendant of tag
-  // 'noscript'. Did you mean 'amp-img'?"
-  if (error.code === amp.validator.ValidationError.Code.DISALLOWED_TAG) {
-    if (error.params[0] === 'img' || error.params[0] === 'video' ||
-        error.params[0] === 'audio' || error.params[0] === 'iframe' ||
-        error.params[0] === 'font') {
-      return amp.validator.ErrorCategory.Code
-          .DISALLOWED_HTML_WITH_AMP_EQUIVALENT;
-    }
-    // E.g. "The tag 'picture' is disallowed."
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // E.g. "tag 'img' may only appear as a descendant of tag
-  // 'noscript'. Did you mean 'amp-img'?"
-  if (error.code ===
-      amp.validator.ValidationError.Code.MANDATORY_TAG_ANCESTOR_WITH_HINT) {
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML_WITH_AMP_EQUIVALENT;
-  }
-  // At the moment it's not possible to get this particular error since
-  // all mandatory tag ancestors have hints except for noscript, but
-  // usually when noscript fails then it reports an error for mandatory_parent
-  // (since there is such a TagSpec as well, for the head).
-  if (error.code ===
-      amp.validator.ValidationError.Code.MANDATORY_TAG_ANCESTOR) {
-    if (goog.string./*OK*/ startsWith(error.params[0], 'amp-') ||
-        goog.string./*OK*/ startsWith(error.params[1], 'amp-')) {
-      return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-    }
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // E.g. "Tag 'amp-accordion > section' must have 2 child tags - saw
-  // 3 child tags."
-  if (error.code ==
-      amp.validator.ValidationError.Code.INCORRECT_NUM_CHILD_TAGS) {
-    if (goog.string./*OK*/ startsWith(error.params[0], 'amp-')) {
-      return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-    }
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // e.g. "Tag 'div' is disallowed as first child of tag
-  // 'amp-accordion > section'. Allowed first child tag names are
-  // ['h1', 'h2', 'h3', 'h4', 'h5', 'h6']."
-  if (error.code ==
-          amp.validator.ValidationError.Code.DISALLOWED_CHILD_TAG_NAME ||
-      error.code ==
-          amp.validator.ValidationError.Code.DISALLOWED_FIRST_CHILD_TAG_NAME) {
-    if (goog.string./*OK*/ startsWith(error.params[0], 'amp-') ||
-        goog.string./*OK*/ startsWith(error.params[1], 'amp-')) {
-      return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-    }
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // E.g. "The text (CDATA) inside tag 'style amp-custom' matches
-  // 'CSS !important', which is disallowed."
-  if (error.code === amp.validator.ValidationError.Code.STYLESHEET_TOO_LONG ||
-      (error.code ===
-           amp.validator.ValidationError.Code.CDATA_VIOLATES_BLACKLIST &&
-       error.params[0] === 'style amp-custom')) {
-    return amp.validator.ErrorCategory.Code.AUTHOR_STYLESHEET_PROBLEM;
-  }
-  // E.g. "CSS syntax error in tag 'style amp-custom' - Invalid Declaration."
-  // TODO(powdercloud): Legacy generic css error code. Remove after 2016-06-01.
-  if (error.code === amp.validator.ValidationError.Code.CSS_SYNTAX &&
-      error.params[0] === 'style amp-custom') {
-    return amp.validator.ErrorCategory.Code.AUTHOR_STYLESHEET_PROBLEM;
-  }
-  // E.g. "CSS syntax error in tag 'style amp-custom' - unterminated string."
-  if ((error.code ===
-           amp.validator.ValidationError.Code
-               .CSS_SYNTAX_STRAY_TRAILING_BACKSLASH ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_UNTERMINATED_COMMENT ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_UNTERMINATED_STRING ||
-       error.code === amp.validator.ValidationError.Code.CSS_SYNTAX_BAD_URL ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .CSS_SYNTAX_EOF_IN_PRELUDE_OF_QUALIFIED_RULE ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_DECLARATION ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .CSS_SYNTAX_INCOMPLETE_DECLARATION ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_AT_RULE ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .CSS_SYNTAX_ERROR_IN_PSEUDO_SELECTOR ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_MISSING_SELECTOR ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_NOT_A_SELECTOR_START ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .CSS_SYNTAX_UNPARSED_INPUT_REMAINS_IN_SELECTOR ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_MISSING_URL ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_URL ||
-       error.code ===
-           amp.validator.ValidationError.Code.CSS_SYNTAX_INVALID_URL_PROTOCOL ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .CSS_SYNTAX_DISALLOWED_RELATIVE_URL) &&
-      error.params[0] === 'style amp-custom') {
-    return amp.validator.ErrorCategory.Code.AUTHOR_STYLESHEET_PROBLEM;
-  }
-  // E.g. "The mandatory tag 'boilerplate (noscript)' is missing or
-  // incorrect."
-  if (error.code === amp.validator.ValidationError.Code.MANDATORY_TAG_MISSING ||
-      (error.code ===
-           amp.validator.ValidationError.Code.MANDATORY_ATTR_MISSING &&
-       error.params[0] === '⚡') ||
-      (error.code ===
-           amp.validator.ValidationError.Code
-               .MANDATORY_CDATA_MISSING_OR_INCORRECT &&
-       (goog.string./*OK*/ startsWith(
-            error.params[0], 'head > style : boilerplate') ||
-        goog.string./*OK*/ startsWith(
-            error.params[0], 'noscript > style : boilerplate')))) {
-    return amp.validator.ErrorCategory.Code
-        .MANDATORY_AMP_TAG_MISSING_OR_INCORRECT;
-  }
-  // E.g. "The mandatory tag 'meta name=viewport' is missing or
-  // incorrect."
-  if ((error.code ===
-           amp.validator.ValidationError.Code
-               .DISALLOWED_PROPERTY_IN_ATTR_VALUE ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .INVALID_PROPERTY_VALUE_IN_ATTR_VALUE ||
-       error.code ===
-           amp.validator.ValidationError.Code
-               .MANDATORY_PROPERTY_MISSING_FROM_ATTR_VALUE) &&
-      error.params[2] === 'meta name=viewport') {
-    return amp.validator.ErrorCategory.Code
-        .MANDATORY_AMP_TAG_MISSING_OR_INCORRECT;
-  }
-  // E.g. "The mandatory attribute 'height' is missing in tag 'amp-img'."
-  if (error.code ===
-          amp.validator.ValidationError.Code.ATTR_VALUE_REQUIRED_BY_LAYOUT ||
-      error.code ===
-          amp.validator.ValidationError.Code.IMPLIED_LAYOUT_INVALID ||
-      error.code ===
-          amp.validator.ValidationError.Code.SPECIFIED_LAYOUT_INVALID ||
-      (error.code ===
-       amp.validator.ValidationError.Code
-           .INCONSISTENT_UNITS_FOR_WIDTH_AND_HEIGHT) ||
-      ((error.code === amp.validator.ValidationError.Code.INVALID_ATTR_VALUE ||
-        error.code ===
-            amp.validator.ValidationError.Code.MANDATORY_ATTR_MISSING) &&
-       (error.params[0] === 'width' || error.params[0] === 'height' ||
-        error.params[0] === 'layout'))) {
-    return amp.validator.ErrorCategory.Code.AMP_LAYOUT_PROBLEM;
-  }
-  if (error.code ===
-          amp.validator.ValidationError.Code
-              .ATTR_DISALLOWED_BY_IMPLIED_LAYOUT ||
-      error.code ===
-          amp.validator.ValidationError.Code
-              .ATTR_DISALLOWED_BY_SPECIFIED_LAYOUT) {
-    return amp.validator.ErrorCategory.Code.AMP_LAYOUT_PROBLEM;
-  }
-  // E.g. "The attribute 'src' in tag 'amphtml engine v0.js script'
-  // is set to the invalid value
-  // '//static.breakingnews.com/ads/gptLoader.js'."
-  if (error.code === amp.validator.ValidationError.Code.INVALID_ATTR_VALUE &&
-      error.params[0] === 'src' &&
-      goog.string./*OK*/ endsWith(error.params[1], 'script')) {
-    return amp.validator.ErrorCategory.Code.CUSTOM_JAVASCRIPT_DISALLOWED;
-  }
-  // E.g. "The tag 'script' is disallowed except in specific forms."
-  if (error.code ===
-          amp.validator.ValidationError.Code.GENERAL_DISALLOWED_TAG &&
-      error.params[0] === 'script') {
-    return amp.validator.ErrorCategory.Code.CUSTOM_JAVASCRIPT_DISALLOWED;
-  }
-  // E.g.: "The attribute 'type' in tag 'script type=application/ld+json'
-  // is set to the invalid value 'text/javascript'."
-  if (error.code === amp.validator.ValidationError.Code.INVALID_ATTR_VALUE &&
-      goog.string./*OK*/ startsWith(error.params[1], 'script') &&
-      error.params[0] === 'type') {
-    return amp.validator.ErrorCategory.Code.CUSTOM_JAVASCRIPT_DISALLOWED;
-  }
-  // E.g. "The attribute 'srcset' may not appear in tag 'amp-audio >
-  // source'."
-  if ((error.code === amp.validator.ValidationError.Code.INVALID_ATTR_VALUE ||
-       error.code === amp.validator.ValidationError.Code.DISALLOWED_ATTR ||
-       error.code ===
-           amp.validator.ValidationError.Code.MANDATORY_ATTR_MISSING)) {
-    if (goog.string./*OK*/ startsWith(error.params[1], 'amp-')) {
-      return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-    }
-    if (goog.string./*OK*/ startsWith(error.params[1], 'on')) {
-      return amp.validator.ErrorCategory.Code.CUSTOM_JAVASCRIPT_DISALLOWED;
-    }
-    if (error.params[1] === 'style' ||
-        error.params[1] === 'link rel=stylesheet for fonts') {
-      return amp.validator.ErrorCategory.Code.AUTHOR_STYLESHEET_PROBLEM;
-    }
-    // E.g. "The attribute 'async' may not appear in tag 'link
-    // rel=stylesheet for fonts'."
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // Like the previous example but the tag is params[0] here. This
-  // error should always be for AMP elements thus far, so we don't
-  // check for params[0].
-  if (error.code ===
-      amp.validator.ValidationError.Code.MANDATORY_ONEOF_ATTR_MISSING) {
-    return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-  }
-  // E.g. "The attribute 'shortcode' in tag 'amp-instagram' is deprecated -
-  // use 'data-shortcode' instead."
-  if (error.code === amp.validator.ValidationError.Code.DEPRECATED_ATTR ||
-      error.code === amp.validator.ValidationError.Code.DEPRECATED_TAG) {
-    return amp.validator.ErrorCategory.Code.DEPRECATION;
-  }
-  // E.g. "The parent tag of tag 'source' is 'picture', but it can
-  // only be 'amp-audio'."
-  if (error.code === amp.validator.ValidationError.Code.WRONG_PARENT_TAG) {
-    if (goog.string./*OK*/ startsWith(error.params[0], 'amp-') ||
-        goog.string./*OK*/ startsWith(error.params[1], 'amp-') ||
-        goog.string./*OK*/ startsWith(error.params[2], 'amp-')) {
-      return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-    }
-    // E.g. "The parent tag of tag 'script' is 'body', but it can only
-    // be 'head'".
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // E.g. "The 'amp-image-lightbox extension .js script' tag is
-  // missing or incorrect, but required by 'amp-image-lightbox'."
-  if (error.code ===
-          amp.validator.ValidationError.Code.TAG_REQUIRED_BY_MISSING &&
-      (goog.string./*OK*/ startsWith(error.params[1], 'amp-') ||
-       error.params[1] === 'template')) {
-    return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-  }
-  // E.g. "The attribute 'role' in tag 'amp-img' is missing or incorrect,
-  // but required by attribute 'on'."
-  if (error.code ===
-      amp.validator.ValidationError.Code.ATTR_REQUIRED_BUT_MISSING) {
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // E.g. "Mutually exclusive attributes encountered in tag
-  // 'amp-youtube' - pick one of ['src', 'data-videoid']."
-  if (error.code ===
-          amp.validator.ValidationError.Code.MUTUALLY_EXCLUSIVE_ATTRS &&
-      goog.string./*OK*/ startsWith(error.params[0], 'amp-')) {
-    return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-  }
-  // E.g. "The tag 'boilerplate (noscript) - old variant' appears
-  // more than once in the document."
-  if (error.code === amp.validator.ValidationError.Code.DUPLICATE_UNIQUE_TAG) {
-    return amp.validator.ErrorCategory.Code
-        .MANDATORY_AMP_TAG_MISSING_OR_INCORRECT;
-  }
-  // E.g. "Mustache template syntax in attribute name
-  // 'data-{{&notallowed}}' in tag 'p'."
-  if (error.code ===
-          amp.validator.ValidationError.Code.UNESCAPED_TEMPLATE_IN_ATTR_VALUE ||
-      error.code ===
-          amp.validator.ValidationError.Code.TEMPLATE_PARTIAL_IN_ATTR_VALUE ||
-      error.code === amp.validator.ValidationError.Code.TEMPLATE_IN_ATTR_NAME) {
-    return amp.validator.ErrorCategory.Code.AMP_HTML_TEMPLATE_PROBLEM;
-  }
-  // E.g. "The tag 'amp-ad' may not appear as a descendant of tag 'amp-sidebar'.
-  if (error.code ===
-          amp.validator.ValidationError.Code.DISALLOWED_TAG_ANCESTOR &&
-      (goog.string./*OK*/ startsWith(error.params[1], 'amp-'))) {
-    return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-  }
-  if (error.code ===
-          amp.validator.ValidationError.Code.DISALLOWED_TAG_ANCESTOR &&
-      (error.params[1] === 'template')) {
-    return amp.validator.ErrorCategory.Code.AMP_HTML_TEMPLATE_PROBLEM;
-  }
-  // E.g. "Missing URL for attribute 'href' in tag 'a'."
-  // E.g. "Invalid URL protocol 'http:' for attribute 'src' in tag
-  // 'amp-iframe'." Note: Parameters in the format strings appear out
-  // of order so that error.params(1) is the tag for all four of these.
-  if (error.code == amp.validator.ValidationError.Code.MISSING_URL ||
-      error.code == amp.validator.ValidationError.Code.INVALID_URL ||
-      error.code == amp.validator.ValidationError.Code.INVALID_URL_PROTOCOL ||
-      error.code ==
-          amp.validator.ValidationError.Code.DISALLOWED_RELATIVE_URL) {
-    if (goog.string./*OK*/ startsWith(error.params[1], 'amp-')) {
-      return amp.validator.ErrorCategory.Code.AMP_TAG_PROBLEM;
-    }
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  // E.g. "The dimension '1x' in attribute 'srcset' appears more than once."
-  if (error.code == amp.validator.ValidationError.Code.DUPLICATE_DIMENSION) {
-    return amp.validator.ErrorCategory.Code.DISALLOWED_HTML;
-  }
-  return amp.validator.ErrorCategory.Code.GENERIC;
-};
-
-/**
- * Convenience function which calls |CategorizeError| for each error
- * in |result| and sets its category field accordingly.
- * @param {!amp.validator.ValidationResult} result
- * @export
- */
-amp.validator.annotateWithErrorCategories = function(result) {
-  for (const error of result.errors) {
-    error.category = amp.validator.categorizeError(error);
-  }
-};
 
 /**
  * Convenience function which informs caller if given ValidationError is
